+++ plotpy/widgets/labels.py
@@ -105,8 +105,8 @@ class RotatedLabel(QW.QLabel):
         hint = QW.QLabel.sizeHint(self)
         width, height = hint.width(), hint.height()
         angle = self.angle * pi / 180
-        rotated_width = abs(width * cos(angle)) + abs(height * sin(angle))
-        rotated_height = abs(width * sin(angle)) + abs(height * cos(angle))
+        rotated_width = int(abs(width * cos(angle)) + abs(height * sin(angle)))
+        rotated_height = int(abs(width * sin(angle)) + abs(height * cos(angle)))
         return QC.QSize(rotated_width, rotated_height)
 
     def minimumSizeHint(self):
